Revamp Psyche's spellbook
The Orc Wizard III book isn't that exciting. Psyche's defining
characteristics as a unique are her dagger brand and chaos theme, so
I've leaned into that. The new book is: Cantrip, Polymorph, Chain of
Chaos, Invisibility.

At Psyche's depth Chain of Chaos is rather scary, but can be mitigated
by having other monsters around. Since she arrives at a depth where many
characters won't have a ready source of allies, this puts pressure to
engage Psyche with other hostiles around.

Polymorph is also in the mix, partly to thin out the CoC terror with
something chaotic in the casting weight, and partly because polymorph
aimed at the player is a rather rare effect and a good "signature spell"
for a chaos unique.