New spell: Projected Weapon (L5 Translocations)
Projected Weapon launches attacks with your melee weapon against 2-4
random enemies in LOS, depending on spellpower. It's themed as opening
one "in" portal and a bunch of tiny "out" portals, letting you hit many
enemies with a single attack.

(More exciting name wanted!)

This spell is intended to give melee-focused characters more reason to
consider investing in spellcasting, It's fun to try to skill toward a
'hybrid' build, and a spell that mostly scales off your melee attack,
rather than your spellpower, helps encourage that. This being a L5
spell helps avoid it being a truly free 'gimme', and allows it to hit
many enemies at once.

The lack of targeter should hopefully avoid it being annoying to use,
by making it just a single press to cast. It also gives the player
more reason to keep normally hitting things in melee sometimes,
when it's important to kill a specific target.

UC attacks are banned to avoid this being too cheap for transmuters,
which already have light armour and spellcasting skills. That could be
the wrong call!

Cleaving also doesn't work with this spell, since it seems overly
confusing, and since reach and riposte don't work with it either.
(Thus, it's 'fair'.)